Region loss multi: enable GPU computation.
Note, however, that computing the loss on the GPU seems to be slower (probably due to extensive looping instead of batched computations). Thus computing the loss on the CPU is still the default. The loss is by far the slowest part of the training - it takes ~60x as long as a forward pass. There's probably room for improvement with a good CUDA implementation, but I think it would be non-trivial.

I also cleaned up the code a bit and made it just a little faster on CPU.